[BACKEND] Multithread support added.
This feature adds a fully new support to multithreading. It means that now many clients may connect to the server and have the same usage experience. This update comes with the need for multiclient support due future changes that also involve shifting to UDP protocol on the transport layer instead of TCP.

changelog: feature